refactor: #25 - 이벤트 list -> 이벤트 로 레이어 분리
